proto/pyproject.toml
[project]
name = "fruit-ninja-cv"
version = "0.1.0"
description = "A computer vision-based Fruit Ninja game using hand gesture recognition"
readme = "README.md"
requires-python = ">=3.11"
authors = [
    {name = "bigalex95"}
]
keywords = ["game", "computer-vision", "hand-tracking", "mediapipe", "opencv"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
]

dependencies = [
    "mediapipe>=0.10.14",
    "opencv-python>=4.12.0.88",
    "supabase>=2.0.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
fruit-ninja = "src.main:main"

[project.urls]
Homepage = "https://github.com/bigalex95/minininja"
Repository = "https://github.com/bigalex95/minininja"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]
